Moves greet and farwell to display
