+++ src/main/java/com/samhan/ui/Display.java
@@ -5,4 +5,7 @@ import com.samhan.Marker;
 
 public interface Display {
     void render(Board board, Marker marker);
+    void greet();
+    void farewell();
+
 }
